requirements.txt
farcaster-py==0.0.8
blake3
pynacl
